+++ tpi-reportes-service/src/main/java/com/tpi/reportes/dtos/EmpleadoDTO.java
@@ -1,0 +1,38 @@
+package com.tpi.reportes.dtos;
+
+public class EmpleadoDTO {
+    private Long legajo;
+    private String nombre;
+    private String apellido;
+
+    public EmpleadoDTO() {
+    }
+
+    public Long getLegajo() {
+        return legajo;
+    }
+
+    public void setLegajo(Long legajo) {
+        this.legajo = legajo;
+    }
+
+    public String getNombre() {
+        return nombre;
+    }
+
+    public void setNombre(String nombre) {
+        this.nombre = nombre;
+    }
+
+    public String getApellido() {
+        return apellido;
+    }
+
+    public void setApellido(String apellido) {
+        this.apellido = apellido;
+    }
+
+    public String getNombreCompleto() {
+        return nombre + " " + apellido;
+    }
+}
