+++ tpi-reportes-service/src/main/java/com/tpi/reportes/dtos/InteresadoDTO.java
@@ -1,0 +1,29 @@
+package com.tpi.reportes.dtos;
+
+public class InteresadoDTO {
+    private String nombre;
+    private String apellido;
+
+    public InteresadoDTO() {
+    }
+
+    public String getNombre() {
+        return nombre;
+    }
+
+    public void setNombre(String nombre) {
+        this.nombre = nombre;
+    }
+
+    public String getApellido() {
+        return apellido;
+    }
+
+    public void setApellido(String apellido) {
+        this.apellido = apellido;
+    }
+
+    public String getNombreCompleto() {
+        return nombre + " " + apellido;
+    }
+}
