+++ tpi-reportes-service/src/main/java/com/tpi/reportes/dtos/PruebaDTO.java
@@ -79,83 +79,3 @@ public class PruebaDTO {
         this.interesado = interesado;
     }
 }
-public class VehiculoDTO{
-    private Integer id;
-    private String patente;
-    public VehiculoDTO(){}
-
-    public Integer getId() {
-        return id;
-    }
-
-    public void setId(Integer id) {
-        this.id = id;
-    }
-
-    public String getPatente() {
-        return patente;
-    }
-
-    public void setPatente(String patente) {
-        this.patente = patente;
-    }
-}
-public class EmpleadoDTO{
-    private Long legajo;
-    private String nombre;
-    private String apellido;
-
-    public EmpleadoDTO(){}
-
-    public Long getLegajo() {
-        return legajo;
-    }
-
-    public void setLegajo(Long legajo) {
-        this.legajo = legajo;
-    }
-
-    public String getNombre() {
-        return nombre;
-    }
-
-    public void setNombre(String nombre) {
-        this.nombre = nombre;
-    }
-
-    public String getApellido() {
-        return apellido;
-    }
-
-    public void setApellido(String apellido) {
-        this.apellido = apellido;
-    }
-    public String getNombreCompleto(){
-        return  nombre + " " + apellido;
-    }
-}
-public class InteresadoDTO{
-    private String nombre;
-    private String apellido;
-
-    public InteresadoDTO(){}
-
-    public String getNombre() {
-        return nombre;
-    }
-
-    public void setNombre(String nombre) {
-        this.nombre = nombre;
-    }
-
-    public String getApellido() {
-        return apellido;
-    }
-
-    public void setApellido(String apellido) {
-        this.apellido = apellido;
-    }
-    public String getNombreCompleto(){
-        return  nombre + " " + apellido;
-    }
-}
